requirements.txt
xlwings[all]
fintoolkit